# RANK: Optimize OTSU filter