--- a/skimage/filters/rank/generic_cy.pyx
+++ b/skimage/filters/rank/generic_cy.pyx
@@ -385,24 +385,27 @@ cdef inline void _kernel_otsu(dtype_t_out* out, Py_ssize_t odepth,
     if pop:
         for i in range(n_bins):
             mu += histo[i] * i
-        mu = mu / pop
     else:
         out[0] = <dtype_t_out>0
+        return
+
 
     # maximizing the between class variance
     max_i = 0
-    q1 = histo[0] / pop
+    q1 = histo[0]
     mu1 = 0.
     max_sigma_b = 0.
 
     for i in range(1, n_bins):
-        P = histo[i] / pop
+        P = histo[i]
         new_q1 = q1 + P
+        if new_q1 == pop:
+            break
         if new_q1 > 0:
-            mu1 = (q1 * mu1 + i * P) / new_q1
-            mu2 = (mu - new_q1 * mu1) / (1. - new_q1)
-            t = mu1 - mu2
-            sigma_b = new_q1 * (1. - new_q1) * (t * t)
+            mu1 = mu1 + i * P
+            mu2 = mu - mu1
+            t = (pop - new_q1) * mu1 - mu2 * new_q1
+            sigma_b = (t * t) / (new_q1 * (pop - new_q1))
             if sigma_b > max_sigma_b:
                 max_sigma_b = sigma_b
                 max_i = i