@pep8speaks 2018-10-23
Hello @hmaarrfk! Thanks for submitting the PR.

 - In the file [`benchmarks/benchmark_rank.py`](https://github.com/hmaarrfk/scikit-image/blob/4752fa58172e18706dd10831f037850a7893615d/benchmarks/benchmark_rank.py), following                are the PEP8 issues :
> [Line 14:5](https://github.com/hmaarrfk/scikit-image/blob/4752fa58172e18706dd10831f037850a7893615d/benchmarks/benchmark_rank.py#L14): [E303](https://duckduckgo.com/?q=pep8%20E303) too many blank lines (2)