@hmaarrfk 2018-11-03
@ahojnnes I scaled back the changes not included the `double pop` -> `integer pop`. Honestly, I think using double is a mistake, but it is probably best left to an other PR.