@hmaarrfk 2018-11-05
`otsu~4` is like master. The one that doesn't change was due to the double->Py_ssize_t change that I reverted.  

Master is about (2ms, 120 ms), this PR is at (500us, 25ms) for small and large images.
To get asv to output this, you need to do:

1. Delete the gigantic matrix of builds that are currently benched otherwise it will take for ever.
2. `asv run master..head -b RankSuite`

In this case, it is going to to build 4 or 5 envs, and run them in **slow** mode.

Fast mode isn't super reliable sometime. I know many libraries are change averse when the output changes given a certain input, so I wanted to make sure I portrayed the gains correctly. On my computer, the output seems to have changed now for an other input that happaned to be a pathological edge case.