@jni 2018-11-06
@hmaarrfk good stuff!

I don't particularly want fast mode, I just want to use my current environment by default. Perhaps a PR to asv is warranted...