@hmaarrfk 2018-11-06
yeah, using -E existing is super useful. But sometimes doing a real benchmark is useful. I'm torn on benchmarks.